feat: meal amount mapper, controller

+++ src/main/java/com/example/let/server/domain/mealamount/dto/request/MealAmountRequest.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.NotBlank;
 
 public record MealAmountRequest(
     @NotBlank
-    Long mealId,
+    Integer mealId,
     @NotBlank
     Rating rating
 ) {
